# lib/api-zod/src/generated/types/itemUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart360 API - multi-tenant guest information PWA
 * OpenAPI spec version: 0.1.0
 */

export interface ItemUpdate {
  categoryId?: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  body?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  priceUnit?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  mapQuery?: string | null;
  /** @nullable */
  difficulty?: string | null;
  /** @nullable */
  duration?: string | null;
  /** @nullable */
  distance?: string | null;
  open24?: boolean;
  /** @nullable */
  hoursJson?: string | null;
  /** @nullable */
  noteType?: string | null;
  /** @nullable */
  noteText?: string | null;
  bullets?: string[];
  position?: number;
  isVisible?: boolean;
}