# lib/api-zod/src/generated/types/mediaEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart360 API - multi-tenant guest information PWA
 * OpenAPI spec version: 0.1.0
 */

export interface MediaEntry {
  id: string;
  /** @nullable */
  itemId?: string | null;
  /** @nullable */
  tenantId?: string | null;
  url: string;
  /** @nullable */
  alt?: string | null;
  position: number;
}